+++ src/Foundation/Providers/EventServiceProvider.php
@@ -15,9 +15,12 @@ declare(strict_types=1);
 namespace Vanilo\Foundation\Providers;
 
 use Illuminate\Foundation\Support\Providers\EventServiceProvider as ServiceProvider;
+use Vanilo\Cart\Events\CartDeleting;
+use Vanilo\Cart\Events\CartUpdated;
 use Vanilo\Checkout\Events\ShippingAddressChanged;
 use Vanilo\Checkout\Events\ShippingMethodSelected;
 use Vanilo\Foundation\Listeners\CalculateShippingFees;
+use Vanilo\Foundation\Listeners\DeleteCartAdjustments;
 use Vanilo\Foundation\Listeners\UpdateSalesFigures;
 use Vanilo\Order\Events\OrderWasCreated;
 
@@ -33,5 +36,11 @@ class EventServiceProvider extends ServiceProvider
         ShippingAddressChanged::class => [
             CalculateShippingFees::class,
         ],
+        CartUpdated::class => [
+            CalculateShippingFees::class,
+        ],
+        CartDeleting::class => [
+            DeleteCartAdjustments::class,
+        ]
     ];
 }