+++ src/Adjustments/Changelog.md
@@ -2,6 +2,12 @@
 
 ## 3.x Series
 
+## Unreleased
+##### 2023-XX-YY
+
+- Added the `clear()` method to `AdjustmentCollection` implementations
+- Added the `invalidateAdjustments()` method to `Adjustable` implementations
+
 ## 3.5.0
 ##### 2023-02-23